requirements/local.txt

# Include the base.txt requirements file
-r base.txt

# Local development dependencies go here
django-debug-toolbar==1.2.2
